requirements.txt
appnope==0.1.4
asgiref==3.7.2
asttokens==2.4.1
bcrypt==4.1.3
boto3==1.34.158
botocore==1.34.158
Brotli==1.1.0
cffi==1.16.0
comm==0.2.2
cryptography==42.0.7
cssselect2==0.7.0
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
dj-database-url==2.1.0
Django==4.1.13
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-livereload-server==0.5.1
django-storages==1.14.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
Fabric3==1.14.post1
fabtools-python==0.19.4
fontawesomefree==6.5.1
fonttools==4.51.0
future==1.0.0
greenlet==3.0.3
gunicorn==22.0.0
html5lib==1.1
importlib_metadata==7.1.0
invoke==2.2.0
ipykernel==6.29.4
ipython==8.18.1
isoweek==1.3.3
jedi==0.19.1
jmespath==1.0.1
joblib==1.4.0
jupyter_client==8.6.1
jupyter_core==5.7.2
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
paramiko==2.12.0
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydyf==0.9.0
Pygments==2.17.2
PyNaCl==1.5.0
pyodbc==5.1.0
pyphen==0.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-logstash==0.4.8
pytz==2024.1
pyzmq==25.1.2
s3transfer==0.10.2
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
SQLAlchemy==2.0.29
sqlparse==0.4.4
stack-data==0.6.3
threadpoolctl==3.4.0
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
weasyprint==61.2
webencodings==0.5.1
whitenoise==6.6.0
wrapt==1.16.0
zipp==3.18.1
zopfli==0.2.3
